pyproject.toml
[project]
name = "zotero-pdf-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "jupyterlab>=4.3.5",
    "lxml>=5.3.1",
    "marker-pdf>=1.6.0",
    "pandas>=2.2.3",
    "pyzotero>=1.6.9",
]
